4-12

+++ realtime/src/main/java/com/gmall/realtime/bean/ValueName.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class ValueName {
-    private Integer value;
+    private double value;
     private String name;
 }
